feat: migrate to typescript
BREAKING CHANGE: Large re-write to TypeScript
