+++ .gitignore
@@ -3,3 +3,4 @@ node_modules/
 .rpt2_cache/
 *.log
 dist/
+.rts2_*
