+++ src/SelectAll.tsx
@@ -1,7 +1,16 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import * as React from 'react';
 import { onlyUpdateForKeys } from 'recompose';
-function SelectAll({
+
+type SelectAllProps = {
+  tabIndex: number | undefined;
+  disabled: boolean;
+  allSelected: boolean;
+  id: string;
+  selectAllText?: string;
+  toggleSelectAll(): void;
+  visible: boolean;
+};
+const SelectAll: React.FC<SelectAllProps> = ({
   tabIndex,
   disabled,
   allSelected,
@@ -9,7 +18,7 @@ function SelectAll({
   selectAllText,
   toggleSelectAll,
   visible,
-}) {
+}) => {
   if (!visible) {
     return null;
   }
@@ -23,7 +32,6 @@ function SelectAll({
       aria-selected={allSelected}
       className={allSelected ? 'option selected' : 'option'}
       onClick={toggleSelectAll}
-      disabled={disabled}
       onKeyPress={toggleSelectAll}
     >
       <input
@@ -38,17 +46,9 @@ function SelectAll({
       <span data-testid="select-all-text">{selectAllText}</span>
     </div>
   );
-}
-
-SelectAll.propTypes = {
-  tabIndex: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-  disabled: PropTypes.bool,
-  allSelected: PropTypes.bool,
-  id: PropTypes.string.isRequired,
-  selectAllText: PropTypes.string,
-  toggleSelectAll: PropTypes.func.isRequired,
-  visible: PropTypes.bool,
 };
+
+SelectAll.displayName = 'Picky(SelectAll)';
 
 export default onlyUpdateForKeys([
   'tabIndex',
