+++ src/index.ts
@@ -1,0 +1,3 @@
+import Picky from './Picky';
+
+export default Picky;
