--- src/lib/debounce.js
@@ -1,11 +1,0 @@
-export default (fn, delay) => {
-  let timeoutID = null;
-  return function() {
-    clearTimeout(timeoutID);
-    const args = arguments;
-    const that = this;
-    timeoutID = setTimeout(() => {
-      fn.apply(that, args);
-    }, delay);
-  };
-};
